Move get_cache_key out of REST handler
